+++ src/decorators/createResolverClassMiddlewareDecorator.ts
@@ -1,7 +1,7 @@
 import { type MiddlewareFn } from "@/typings";
 import { UseMiddleware } from "./UseMiddleware";
 
-export function createClassMiddlewareDecorator<TContextType extends object = object>(
+export function createResolverClassMiddlewareDecorator<TContextType extends object = object>(
   resolver: MiddlewareFn<TContextType>,
 ): ClassDecorator {
   return UseMiddleware(resolver);