+++ src/shim.ts
@@ -47,7 +47,7 @@ export const Authorized: typeof src.Authorized = dummyDecorator;
 export const createParamDecorator: typeof src.createParamDecorator = dummyFn as any;
 export const createMethodMiddlewareDecorator: typeof src.createMethodMiddlewareDecorator =
   dummyFn as any;
-export const createClassMiddlewareDecorator: typeof src.createClassMiddlewareDecorator =
+export const createResolverClassMiddlewareDecorator: typeof src.createResolverClassMiddlewareDecorator =
   dummyFn as any;
 export const Ctx: typeof src.Ctx = dummyDecorator;
 export const Directive: typeof src.Directive = dummyDecorator;