+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - support declaring middlewares on resolver class level (#620)
 - support declaring auth roles on resolver class level (#620)
+- make possible creating custom decorators on resolver class level - `createResolverClassMiddlewareDecorator`
 
 ### Others